congestion-sim/requirements.txt
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
pandas>=2.0
pyyaml>=6.0
jupyter
